JSF - list of all users + functionality to select user, see kweets and delete a kweet

+++ src/java/services/UserService.java
@@ -3,6 +3,7 @@ package services;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.faces.bean.ApplicationScoped;
+import javax.faces.bean.ManagedBean;
 import javax.inject.Inject;
 import javax.persistence.*;
 import java.security.InvalidParameterException;
@@ -36,7 +37,9 @@ public class UserService{
     }
 
     public boolean removeUser(Long id, Long userId){
-
+        if(id ==  userId){
+            return false;
+        }
         return this.userDAO.removeUser(id, userId);
     }
 
@@ -44,7 +47,7 @@ public class UserService{
         if(user.getBiography().length() > 160){
             return false;
         }
-        if(user.getLocation().length() > 160){
+        if(user.getLocation().length() > 100){
             return false;
         }
         return this.userDAO.updateUser(user, id);
@@ -54,20 +57,34 @@ public class UserService{
         return userDAO.findUserById(id);
     }
 
+    public List<User> findUserByUsername(String name){return userDAO.findUserByName(name);}
+
     public boolean addFollowing(Long user, Long following) {
+        if(user == following){
+            return false;
+        }
         return userDAO.addFollowing(user, following);
     }
 
     public boolean addFollower(Long user, Long follower) {
+        if(user == follower){
+            return false;
+        }
         return userDAO.addFollower(user, follower);
     }
 
     public boolean removeFollower(Long user, Long follower) {
-        return userDAO.removeFollower(user, follower);
+        if(user != follower){
+            return userDAO.removeFollower(user, follower);
+        }
+        return false;
     }
 
     public boolean removeFollowing(Long id, Long following){
-        return userDAO.removeFollowing(id, following);
+        if(id != following){
+            return userDAO.removeFollowing(id, following);
+        }
+        return false;
     }
 
     public List<User> getAllUsers(){
